done nest auth with passport

+++ nest-passport-auth/src/user/user.service.ts
@@ -1,11 +1,20 @@
-import { Injectable } from "@nestjs/common";
-import { User } from "@prisma/client";
+import { IJwtPayload } from "@auth/interfaces";
+import { CACHE_MANAGER } from "@nestjs/cache-manager";
+import { ForbiddenException, Inject, Injectable } from "@nestjs/common";
+import { ConfigService } from "@nestjs/config";
+import { Role, User } from "@prisma/client";
 import { PrismaService } from "@prisma/prisma.service";
+import { convertToSecondsUtil } from "@share/utils";
 import { hashSync, genSaltSync } from "bcrypt";
+import { Cache } from "cache-manager";
 
 @Injectable()
 export class UserService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly configService: ConfigService,
+    @Inject(CACHE_MANAGER) private cacheManager: Cache
+  ) {}
 
   async save(user: Partial<User>) {
     const hashedPassword = this.hashPassword(user.password);
@@ -19,18 +28,50 @@ export class UserService {
     });
   }
 
-  async findOneByIdOrEmail(idOrEmail: string) {
-    return await this.prisma.user.findFirst({
-      where: {
-        OR: [{ id: idOrEmail }, { email: idOrEmail }],
-      },
-    });
+  async findOneByIdOrEmail(idOrEmail: string, isReset = false) {
+    if (isReset) {
+      await this.cacheManager.del(idOrEmail);
+    }
+
+    const user = await this.cacheManager.get<User>(idOrEmail);
+
+    if (!user) {
+      const user = await this.prisma.user.findFirst({
+        where: {
+          OR: [{ id: idOrEmail }, { email: idOrEmail }],
+        },
+      });
+
+      if (!user) {
+        return null;
+      }
+
+      await this.cacheManager.set(
+        idOrEmail,
+        user,
+        convertToSecondsUtil(this.configService.get("JWT_EXPIRE"))
+      );
+    }
+
+    return user;
   }
 
-  async delete(id: string) {
+  async delete(id: string, user: IJwtPayload) {
+    if (user.id !== id && !user.roles.includes(Role.Admin)) {
+      throw new ForbiddenException();
+    }
+
+    await Promise.all([
+      this.cacheManager.del(id),
+      this.cacheManager.del(user.email),
+    ]);
+
     return await this.prisma.user.delete({
       where: {
         id,
+      },
+      select: {
+        id: true,
       },
     });
   }
